refactor(transcription): split backends/sherpa.rs into module
Mechanical split with no behavior changes:
- mod.rs: SherpaBackend facade + TranscriptionBackend impl + tests
- host.rs: SHERPA_HOST, spawn, run_host_loop, shared constants
- streaming.rs: build_recognizer_config, run_session, finalize_session

Sets up the Moonshine integration by giving us a clean place for
offline.rs and silero_vad.rs without pushing sherpa.rs over 1000 lines.
